consumer/requirements.txt
confluent-kafka
python-dotenv==1.0.1
psycopg[binary]>=3.1
